+++ examples/vn_monitoring/path_analyzer_test/main.tf
@@ -1,10 +1,6 @@
 // Copyright (c) 2017, 2024, Oracle and/or its affiliates. All rights reserved.
 // Licensed under the Mozilla Public License v2.0
 
-variable "tenancy_ocid" {}
-variable "user_ocid" {}
-variable "fingerprint" {}
-variable "private_key_path" {}
 variable "region" {}
 variable "compartment_ocid" {}
 
@@ -33,7 +29,7 @@ variable "path_analyzer_test_protocol" {
 }
 
 variable "path_analyzer_test_protocol_parameters_destination_port" {
-  default = 10
+  default = 0
 }
 
 variable "path_analyzer_test_protocol_parameters_icmp_code" {
@@ -45,7 +41,7 @@ variable "path_analyzer_test_protocol_parameters_icmp_type" {
 }
 
 variable "path_analyzer_test_protocol_parameters_source_port" {
-  default = 10
+  default = 0
 }
 
 variable "path_analyzer_test_protocol_parameters_type" {
@@ -69,17 +65,107 @@ variable "path_analyzer_test_state" {
 }
 
 
-
+variable "config_file_profile" {
+  default = ""
+}
 provider "oci" {
-  tenancy_ocid     = var.tenancy_ocid
-  user_ocid        = var.user_ocid
-  fingerprint      = var.fingerprint
-  private_key_path = var.private_key_path
   region           = var.region
+  auth = "SecurityToken"
+  config_file_profile = var.config_file_profile
+}
+
+variable "on_prem_path_analyzer_test_destination_endpoint_type" {
+  default = "ON_PREM"
+}
+
+variable "on_prem_path_analyzer_test_destination_endpoint_address" {
+  default = "10.2.41.5"
+}
+
+
+variable "on_prem_path_analyzer_test_protocol" {
+  default = 6
+}
+
+
+variable "on_prem_path_analyzer_test_source_endpoint_type" {
+  default = "IP_ADDRESS"
+}
+
+
+variable "on_prem_path_analyzer_test_source_endpoint_address" {
+  default = "100.130.10.100"
+}
+
+
+variable "on_prem_path_analyzer_test_display_name" {
+  default = "On Premise Path Analyzer Test"
+}
+
+
+variable "on_prem_path_analyzer_test_freeform_tags" {
+  default = { "bar-key" = "value" }
+}
+
+
+variable "on_prem_path_analyzer_test_protocol_parameters_type" {
+  default = "TCP"
+}
+
+
+variable "on_prem_path_analyzer_test_protocol_parameters_destination_port" {
+  default = 85
+}
+
+variable "on_prem_path_analyzer_test_protocol_parameters_source_port" {
+  default = 84
+}
+
+
+variable "on_prem_path_analyzer_test_query_options_is_bi_directional_analysis" {
+  default = false
 }
 
 resource "oci_vn_monitoring_path_analyzer_test" "test_path_analyzer_test" {
   #Required
+  compartment_id = var.compartment_ocid
+  destination_endpoint  {
+    #Required
+    type = var.on_prem_path_analyzer_test_destination_endpoint_type
+
+    #Optional
+    address                  = var.on_prem_path_analyzer_test_destination_endpoint_address
+
+  }
+  protocol = var.on_prem_path_analyzer_test_protocol
+  source_endpoint  {
+    #Required
+    type = var.on_prem_path_analyzer_test_source_endpoint_type
+
+    #Optional
+    address                  = var.on_prem_path_analyzer_test_source_endpoint_address
+  }
+
+  #Optional
+  display_name  = var.on_prem_path_analyzer_test_display_name
+  freeform_tags = var.on_prem_path_analyzer_test_freeform_tags
+  protocol_parameters  {
+    #Required
+    type = var.on_prem_path_analyzer_test_protocol_parameters_type
+
+    #Optional
+    destination_port = var.on_prem_path_analyzer_test_protocol_parameters_destination_port
+    source_port      = var.on_prem_path_analyzer_test_protocol_parameters_source_port
+  }
+  query_options  {
+
+    #Optional
+    is_bi_directional_analysis = var.on_prem_path_analyzer_test_query_options_is_bi_directional_analysis
+  }
+}
+
+resource "oci_vn_monitoring_path_analyzer_test" "test_on_prem_path_analyzer_test" {
+  #Requiredn
   compartment_id = var.compartment_ocid
   destination_endpoint  {
     #Required