+++ internal/service/log_analytics/log_analytics_log_analytics_entity_topology_data_source.go
@@ -21,6 +21,13 @@ func LogAnalyticsLogAnalyticsEntityTopologyDataSource() *schema.Resource {
 				Type:     schema.TypeString,
 				Required: true,
 			},
+			"metadata_equals": {
+				Type:     schema.TypeList,
+				Optional: true,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+				},
+			},
 			"namespace": {
 				Type:     schema.TypeString,
 				Required: true,
@@ -159,6 +166,123 @@ func LogAnalyticsLogAnalyticsEntityTopologyDataSource() *schema.Resource {
 												"timezone_region": {
 													Type:     schema.TypeString,
 													Computed: true,
+													Elem: &schema.Resource{
+														Schema: map[string]*schema.Schema{
+															// Required
+
+															// Optional
+
+															// Computed
+															"are_logs_collected": {
+																Type:     schema.TypeBool,
+																Computed: true,
+															},
+															"cloud_resource_id": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"compartment_id": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"defined_tags": {
+																Type:     schema.TypeMap,
+																Computed: true,
+																Elem:     schema.TypeString,
+															},
+															"entity_type_internal_name": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"entity_type_name": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"freeform_tags": {
+																Type:     schema.TypeMap,
+																Computed: true,
+																Elem:     schema.TypeString,
+															},
+															"id": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"lifecycle_details": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"management_agent_id": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"metadata": {
+																Type:     schema.TypeList,
+																Computed: true,
+																Elem: &schema.Resource{
+																	Schema: map[string]*schema.Schema{
+																		// Required
+
+																		// Optional
+
+																		// Computed
+																		"items": {
+																			Type:     schema.TypeList,
+																			Computed: true,
+																			Elem: &schema.Resource{
+																				Schema: map[string]*schema.Schema{
+																					// Required
+
+																					// Optional
+
+																					// Computed
+																					"name": {
+																						Type:     schema.TypeString,
+																						Computed: true,
+																					},
+																					"type": {
+																						Type:     schema.TypeString,
+																						Computed: true,
+																					},
+																					"value": {
+																						Type:     schema.TypeString,
+																						Computed: true,
+																					},
+																				},
+																			},
+																		},
+																	},
+																},
+															},
+															"name": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"source_id": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"state": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"time_created": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"time_last_discovered": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"time_updated": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+															"timezone_region": {
+																Type:     schema.TypeString,
+																Computed: true,
+															},
+														},
+													},
 												},
 											},
 										},
@@ -197,6 +321,19 @@ func (s *LogAnalyticsLogAnalyticsEntityTopologyDataSourceCrud) Get() error {
 	if logAnalyticsEntityId, ok := s.D.GetOkExists("log_analytics_entity_id"); ok {
 		tmp := logAnalyticsEntityId.(string)
 		request.LogAnalyticsEntityId = &tmp
+	}
+
+	if metadataEquals, ok := s.D.GetOkExists("metadata_equals"); ok {
+		interfaces := metadataEquals.([]interface{})
+		tmp := make([]string, len(interfaces))
+		for i := range interfaces {
+			if interfaces[i] != nil {
+				tmp[i] = interfaces[i].(string)
+			}
+		}
+		if len(tmp) != 0 || s.D.HasChange("metadata_equals") {
+			request.MetadataEquals = tmp
+		}
 	}
 
 	if namespace, ok := s.D.GetOkExists("namespace"); ok {